Set SHELL to /bin/bash -c
This makes the Dockerfile much more readable. But SHELL is not an OCI
v1.0 feature, so when building with podman, --format=docker must be
used.

I don't see a way to make OpenShift do this. Let's find out if it
defaults to docker v2 or whether we have to revert.